add Type, Photo, Comment controllers

+++ src/components/controllers/Category.js
@@ -36,7 +36,7 @@ export function Index() {
     }, [searchParams])
 
     return (
-        <Table onClick={handleClick}
+        <Table className="mw-md" onClick={handleClick}
             columns={[
                 { title: 'Заголовок', class: 'text-center'},
                 { title: 'Посилання', class: 'text-center'},
@@ -68,6 +68,9 @@ export function Editor() {
     }
 
     const handleDelete = async () => {
+        if (category?.image) {
+            return alert('Спершу видаліть фото')
+        }
         await context.api.delete(['категорії', params.id])
         navigate('/категорії/список')
     }
